+++ test/Fixtures/DummyCommand.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace XtreamwayzTest\Expressive\Messenger\Fixtures;
+
+class DummyCommand
+{
+}