+++ CHANGELOG.md
@@ -2,6 +2,34 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
+## 0.3.0 - 2018-08-22
+
+### Added
+
+- [#3](https://github.com/xtreamwayz/expressive-messenger/pull/3) adds message handling middleware which sends messages 
+  to multiple handlers.
+
+- [#3](https://github.com/xtreamwayz/expressive-messenger/pull/3) adds configuration for transports with DNS. This 
+  enables support for all available enqueue transports.
+
+### Changed
+
+- [#3](https://github.com/xtreamwayz/expressive-messenger/pull/3) changes how a message bus and enqueue transport is 
+  created. This comes with a major configuration overhaul.
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- [#3](https://github.com/xtreamwayz/expressive-messenger/pull/3) removes the redis factory class. The preferred way is 
+  to use DNS to configure enqueue transports.
+
+### Fixed
+
+- Nothing.
+
 ## 0.2.1 - 2018-08-02
 
 ### Added